+++ common/src/main/java/com/lazrproductions/lazrslib/client/font/FontUtilities.java
@@ -45,12 +45,6 @@ public class FontUtilities {
             width = Math.max(width, font.width(text) + 10);
         }
 
-        RenderSystem.setShader(GameRenderer::getPositionTexShader);
-        RenderSystem.defaultBlendFunc();
-        RenderSystem.setShaderColor(1.0F, 1.0F, 1.0F, 1.0F);
-        RenderSystem.disableDepthTest();
-        RenderSystem.disableBlend();
-
         for (int i = 0; i < list.size(); i++) {
             String text = list.get(i).getString();
             graphics.drawString(font, text,
@@ -58,7 +52,6 @@ public class FontUtilities {
                     y + ((list.size() / 2) * space + (space * i)),
                     color, renderShadow);
         }
-        RenderSystem.enableDepthTest();
     }
     /**
      * Draw the given list of components to the screen, each on their own line.
@@ -106,18 +99,10 @@ public class FontUtilities {
      * @param renderShadow Whether to draw the text's drop shadow.
      */
     public static void renderLabel(@Nonnull Font font, @Nonnull GuiGraphics graphics, int x, int y, String text, int color, boolean renderShadow) {
-        RenderSystem.setShader(GameRenderer::getPositionTexShader);
-        RenderSystem.defaultBlendFunc();
-        RenderSystem.setShaderColor(1.0F, 1.0F, 1.0F, 1.0F);
-        RenderSystem.disableDepthTest();
-        RenderSystem.disableBlend();
-
         graphics.drawString(font, text,
                 x - font.width(text) / 2,
                 y,
                 color, renderShadow);
-
-        RenderSystem.enableDepthTest();
     }
     /**
      * Draw text to the screen.
@@ -169,16 +154,8 @@ public class FontUtilities {
      * @param renderShadow Whether to draw the text's drop shadow.
      */
     public static void drawParagraph(@Nonnull Font font, @Nonnull GuiGraphics graphics, int x, int y, @Nonnull String text, int maxWidth, int color, boolean renderShadow) {
-        RenderSystem.setShader(GameRenderer::getPositionTexShader);
-        RenderSystem.defaultBlendFunc();
-        RenderSystem.setShaderColor(1.0F, 1.0F, 1.0F, 1.0F);
-        RenderSystem.disableDepthTest();
-        RenderSystem.disableBlend();
-
         FormattedText para = FormattedText.of(text);
         graphics.drawWordWrap(font, para, x, y, maxWidth, color);
-
-        RenderSystem.enableDepthTest();
     }
     /**
      * Draw text to the screen and wrap it to fit within the given blit coordinates.
@@ -228,16 +205,8 @@ public class FontUtilities {
      * @param renderShadow Whether to draw the text's drop shadow.
      */
     public static void drawParagraph(@Nonnull Font font, @Nonnull GuiGraphics graphics, int x, int y, @Nonnull Component component, int maxWidth, int color, boolean renderShadow) {
-        RenderSystem.setShader(GameRenderer::getPositionTexShader);
-        RenderSystem.defaultBlendFunc();
-        RenderSystem.setShaderColor(1.0F, 1.0F, 1.0F, 1.0F);
-        RenderSystem.disableDepthTest();
-        RenderSystem.disableBlend();
-
         FormattedText para = FormattedText.composite(component);
         graphics.drawWordWrap(font, para, x, y, maxWidth, color);
-
-        RenderSystem.enableDepthTest();
     }
     /**
      * Draw a component to the screen and wrap it to fit within the given blit coordinates.
@@ -287,16 +256,8 @@ public class FontUtilities {
      * @param renderShadow Whether to draw the text's drop shadow.
      */
     public static void drawParagraph(@Nonnull Font font, @Nonnull GuiGraphics graphics, int x, int y, @Nonnull List<Component> list, int maxWidth, int color, boolean renderShadow) {
-        RenderSystem.setShader(GameRenderer::getPositionTexShader);
-        RenderSystem.defaultBlendFunc();
-        RenderSystem.setShaderColor(1.0F, 1.0F, 1.0F, 1.0F);
-        RenderSystem.disableDepthTest();
-        RenderSystem.disableBlend();
-
         FormattedText text = FormattedText.composite(list);
         graphics.drawWordWrap(font, text, x, y, maxWidth, color);
-
-        RenderSystem.enableDepthTest();
     }
     /**
      * Draw a list of components to the screen, each on their own line.
@@ -346,15 +307,7 @@ public class FontUtilities {
      * @param renderShadow Whether to draw the text's drop shadow.
      */
     public static void drawText(@Nonnull Font font, @Nonnull GuiGraphics graphics, int x, int y, @Nonnull String text, int color, boolean renderShadow) {
-        RenderSystem.setShader(GameRenderer::getPositionTexShader);
-        RenderSystem.defaultBlendFunc();
-        RenderSystem.setShaderColor(1.0F, 1.0F, 1.0F, 1.0F);
-        RenderSystem.disableDepthTest();
-        RenderSystem.disableBlend();
-
         graphics.drawString(font, text, x, y, color, renderShadow);
-
-        RenderSystem.enableDepthTest();
     }
     /**
      * Draw text to the screen.
@@ -379,15 +332,7 @@ public class FontUtilities {
      * @param renderShadow Whether to draw the text's drop shadow.
      */
     public static void drawText(@Nonnull Font font, @Nonnull GuiGraphics graphics, int x, int y, @Nonnull Component component, int color, boolean renderShadow) {
-        RenderSystem.setShader(GameRenderer::getPositionTexShader);
-        RenderSystem.defaultBlendFunc();
-        RenderSystem.setShaderColor(1.0F, 1.0F, 1.0F, 1.0F);
-        RenderSystem.disableDepthTest();
-        RenderSystem.disableBlend();
-
         graphics.drawString(font, component, x, y, color, renderShadow);
-
-        RenderSystem.enableDepthTest();
     }
     /**
      * Draw text to the screen.
@@ -411,16 +356,7 @@ public class FontUtilities {
      */
     public static void drawText(@Nonnull Font font, @Nonnull GuiGraphics graphics, @Nonnull ScreenCoordinate pos, @Nonnull String text, int color, boolean renderShadow) {
         pos = pos.withHeight(font.lineHeight).withWidth(font.width(text));
-
-        RenderSystem.setShader(GameRenderer::getPositionTexShader);
-        RenderSystem.defaultBlendFunc();
-        RenderSystem.setShaderColor(1.0F, 1.0F, 1.0F, 1.0F);
-        RenderSystem.disableDepthTest();
-        RenderSystem.disableBlend();
-
         graphics.drawString(font, text, pos.getX(), pos.getY(), color, renderShadow);
-
-        RenderSystem.enableDepthTest();
     }
     /**
      * Draw text to the screen.
@@ -444,16 +380,7 @@ public class FontUtilities {
      */
     public static void drawText(@Nonnull Font font, @Nonnull GuiGraphics graphics, @Nonnull ScreenCoordinate pos, @Nonnull Component component, int color, boolean renderShadow) {
         pos = pos.withHeight(font.lineHeight).withWidth(font.width(component));
-
-        RenderSystem.setShader(GameRenderer::getPositionTexShader);
-        RenderSystem.defaultBlendFunc();
-        RenderSystem.setShaderColor(1.0F, 1.0F, 1.0F, 1.0F);
-        RenderSystem.disableDepthTest();
-        RenderSystem.disableBlend();
-
         graphics.drawString(font, component, pos.getX(), pos.getY(), color, renderShadow);
-
-        RenderSystem.enableDepthTest();
     }