+++ common/src/main/java/com/lazrproductions/lazrslib/client/screen/base/GenericScreen.java
@@ -54,7 +54,5 @@ public abstract class GenericScreen extends Screen {
         lastKeyInput = new InputAction(keyCode, action);
     }
 
-    public void onClose() {
-
-    }
+    public void onClose() {}
 }