+++ forge/src/main/java/com/lazrproductions/lazrslib/LazrsLibMod.java
@@ -2,6 +2,7 @@ package com.lazrproductions.lazrslib;
 
 import net.minecraftforge.api.distmarker.Dist;
 import net.minecraftforge.fml.DistExecutor;
+import net.minecraftforge.fml.ModLoadingContext;
 import net.minecraftforge.fml.common.Mod;
 
 @Mod(LazrsLibConstants.MOD_ID)
@@ -9,5 +10,6 @@ public class LazrsLibMod {
     
     public LazrsLibMod() {
         DistExecutor.safeRunWhenOn(Dist.CLIENT, () -> LazrsLibClientMod::new);
+        var context = ModLoadingContext.get();
     }
 }