+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.2.1] - 2026-08-18
+
+### Added
+- Added several utility methods to `InputAction` for checks and consuming.
+
+### Fixed
+
+- Removed render shader overrides when drawing from several functions in `ScreenUtilities`
+- Removed render shader overrides when drawing from several functions in `FontUtilities`
+
+
 ## [1.2.0] - 2026-08-09
 
 ### Added